Updated for specific case for wasm
