+++ file/file_other.go
@@ -1,20 +1,11 @@
-//go:build !windows && !darwin && !dragonfly && !freebsd && !linux && !nacl && !netbsd && !openbsd && !solaris
+//go:build !windows && !darwin && !dragonfly && !freebsd && !linux && !nacl && !netbsd && !openbsd && !solaris && !wasm
 
 package file
 
 import (
 	"os"
-	"syscall"
 )
 
-func getOSFileInfo(info os.FileInfo) *FileInfo {
-	fi := &FileInfo{}
-	if s, ok := info.Sys().(*syscall.Stat_t); ok {
-		fi.Nlink = uint32(s.Nlink)
-		fi.UID = s.Uid
-		fi.GID = s.Gid
-		fi.Fileid = s.Ino
-		return fi
-	}
+func getOSFileInfo(_ os.FileInfo) *FileInfo {
 	return nil
 }
